+++ src/module/mountManager.ts
@@ -1,11 +1,5 @@
 import { SettingsForm } from './settings-form';
-import {
-  detachAllFromTokenTA,
-  dismountDropAllTA,
-  dismountDropTargetTA,
-  mountUpTA,
-  moveToken,
-} from './tokenAttacherHelper';
+import { dismountDropAllTA, dismountDropTargetTA, mountUpTA } from './tokenAttacherHelper';
 import { findTokenById, MountUpFlags, getTokenCenter, riderLock, riderX, riderY, socketAction } from './utils';
 import { error, log, warn } from './lib/lib';
 import CONSTANTS from './constants';
@@ -211,6 +205,14 @@ export class MountManager {
     await riderToken.document.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Mount);
     await riderToken.document.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.OrigSize);
     await riderToken.document.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.AlreadyMounted);
+
+    // ADDED 2022-06-06
+    await riderToken.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.MountMove);
+    await riderToken.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.OrigElevation);
+
+    if ((<any[]>mountToken.actor?.getFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders))?.length <= 0) {
+      await mountToken.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders);
+    }
     // MOD 4535992 FORCE SHRINK TO OTHERS RIDERS
     //let riders = <string[]>mountToken.actor.getFlag(CONSTANTS.MODULE_NAME, Flags.Riders);
     // for (const riderTmp of riders) {
@@ -292,13 +294,16 @@ export class MountManager {
         const mount = findTokenById(mountId);
         // MOD 4535992 CHECK IF TOKEN IS ALREADY DELETED
         if (mount) {
-          await mount.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders);
-          // TODO to remove
-          await mount.document.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders);
+          MountManager.doRemoveMount(token, mount);
+          // await mount.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders);
+          // // TODO to remove
+          // await mount.document.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders);
         }
       }
 
       if (this.isaMount(token.id)) {
+        MountManager.removeAllRiders(token);
+        /*
         const riders =
           <string[]>token.actor?.getFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Riders) ||
           // TODO to remove
@@ -307,13 +312,14 @@ export class MountManager {
           const rider: Token = findTokenById(riderTmp);
           // MOD 4535992 CHECK IF TOKEN IS ALREADY DELETED
           if (rider) {
-            await rider.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Mount);
-            await rider.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.OrigSize);
-            // TODO to remove
-            await rider.document?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Mount);
-            await rider.document?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.OrigSize);
+            // await rider.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Mount);
+            // await rider.actor?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.OrigSize);
+            // // TODO to remove
+            // await rider.document?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Mount);
+            // await rider.document?.unsetFlag(CONSTANTS.MODULE_NAME, MountUpFlags.OrigSize);
           }
         }
+        */
       }
 
       return true;
@@ -678,7 +684,7 @@ export class MountManager {
   static isAncestor(childId: string, ancestorId: string) {
     if (this.isaRider(childId)) {
       const child: Token = findTokenById(childId);
-      if(!child){
+      if (!child) {
         warn(`No child found on 'isAncestor' for id '${childId}' for ancestor '${ancestorId}'`, true);
         return false;
       }
@@ -687,7 +693,7 @@ export class MountManager {
         // TODO to remove
         <string>child.document.getFlag(CONSTANTS.MODULE_NAME, MountUpFlags.Mount);
       const parent: Token = findTokenById(parentId);
-      if(!parent){
+      if (!parent) {
         warn(`No parent found on 'isAncestor' for id '${parentId}' for ancestor '${ancestorId}'`, true);
         return false;
       }